Create README.md
Your assignment this week you will build a CheckingAccount class and associated driver application.  The class should contain appropriate attributes that you would find in a normal checking account:  name, address, accounts number, balance, etc  The balance should be "private" to the class since it should only be changed through appropriate debit/credit methods.  Recall, you can simulate the private aspect using the _ character at the start of the variable name.   The class should be saved in a file called CheckingAccount.py  Build a driver application that instantiates a CheckingAccount object and performs a variety of debits and credits as well as prints the balance as appropriate.